refactor: Move V8StackTraceIterator to separate files
Extracted the V8StackTraceIterator class from bindings.cpp into
dedicated V8StackTraceIterator.h and V8StackTraceIterator.cpp files.

Changes:
- Created V8StackTraceIterator.h with class declaration
- Created V8StackTraceIterator.cpp with implementation
- Moved class into Bun namespace
- Updated bindings.cpp to include the new header
- Removed class definition and helper function from bindings.cpp
